refactor(server): Move JsonSchemaValidator to zally-core #560
