+++ server/src/main/java/de/zalando/zally/rule/zalando/UseProblemJsonRule.kt
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.google.common.io.Resources
 import de.zalando.zally.core.EMPTY_JSON_POINTER
 import de.zalando.zally.core.plus
-import de.zalando.zally.rule.JsonSchemaValidator
+import de.zalando.zally.core.JsonSchemaValidator
 import de.zalando.zally.core.ObjectTreeReader
 import de.zalando.zally.rule.api.Check
 import de.zalando.zally.rule.api.Context
