+++ server/zally-core/src/main/resources/schema-validation-messages.properties
