+++ server/zally-core/src/main/resources/schemas/json-schema.json
