+++ server/zally-core/src/main/resources/schemas/openapi-2-schema.json
