+++ server/zally-core/src/main/resources/schemas/openapi-3-schema.json
