+++ server/zally-core/src/main/resources/schemas/problem-schema.json
