+++ server/zally-core/src/main/resources/schemas/swagger-schema.json
