+++ server/zally-core/src/test/kotlin/de/zalando/zally/core/JsonSchemaValidatorTest.kt
@@ -1,7 +1,6 @@
-package de.zalando.zally.rule
+package de.zalando.zally.core
 
 import com.google.common.io.Resources
-import de.zalando.zally.core.ObjectTreeReader
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
