+++ server/zally-core/src/test/resources/schemas/simple-referenced-schema.json
