+++ server/zally-core/src/test/resources/schemas/simple-schema.json
